+++ crates/scripts/src/signer.rs
@@ -17,7 +17,7 @@ const MAX_CYCLES: u64 = 1_000_000;
 const PROGRAM_ELF: &[u8] = b"elf";
 const PROGRAM_ID: &[u8] = b"programID";
 const VM_TYPE: i32 = 0;
-const NONCE: u32 = 1;
+const NONCE: u64 = 1;
 const CONSUMER_ADDR: &str = "0xDB8cFf278adCCF9E9b5da745B44E754fC4EE3C76";
 
 /// Script to generate ABI-encoded responses + signatures for the coprocessor contract tests
